pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smae"
version = "0.1.0"
description = "Socio-Metabolic Analytical Engine — tracking global resource appropriation, displacement, and resistance"
requires-python = ">=3.11"
dependencies = [
    "reportlab>=4.0",
    "httpx>=0.27",
    "pydantic>=2.5",
    "click>=8.1",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
smae = "smae.cli.main:cli"

[tool.setuptools.packages.find]
include = ["smae*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
